sf_mobile_base: move wms components to sf_mobile
