+++ shopfloor_mobile/templates/assets.xml
@@ -22,6 +22,11 @@
                 type="module"
             />
             <script
+                id="script_misc"
+                t-attf-src="/shopfloor_mobile/static/wms/src/components/misc.js?v=#{mod_version}"
+                type="module"
+            />
+            <script
                 id="script_component_packaging_qty_picker"
                 t-attf-src="/shopfloor_mobile/static/wms/src/components/packaging-qty-picker.js?v=#{mod_version}"
                 type="module"
